UF-15100: Minor refactoring
